chore: remove meteor.startup from unstar-message
